fix: main.tsx에서 나는 eslint 오류 수정 파일이 React 컴포넌트만을 export 해야 한다는 규칙 위반 네비바 조건 함수는 App.tsx 로 옮김
